implemented shadows

+++ Camera.cpp
@@ -10,17 +10,17 @@ Camera::Camera() : fromPoint(0, 0, -3), atPoint(0, 0, 2), upVector(0, 1, 0), vie
   nearClippingDepth = -fromPoint.z - 0.5f;
   farClippingDepth = -fromPoint.z + 0.5f;
 
-  D = 1 / (2.0f * (float) std::tan(viewingAngle * PI / 180));
+  D = 1 / (2.0f * std::tan(viewingAngle * PI / 180));
 
   computeBasisVectors();
 }
 
 Ray Camera::computeRayAt(int i, int j, int N) {
-  Vector3 pixEyePos = Vector3((float) i / (float) N - 0.5f, (float) j / (float) N - 0.5f, D);
+  Vector3 pixEyePos = Vector3((double) i / (double) N - 0.5f, (double) j / (double) N - 0.5f, D);
 
-  float pixWorldX = b1.dot(pixEyePos) + fromPoint.x;
-  float pixWorldY = b2.dot(pixEyePos) + fromPoint.y;
-  float pixWorldZ = b3.dot(pixEyePos) + fromPoint.z;
+  double pixWorldX = b1.dot(pixEyePos) + fromPoint.x;
+  double pixWorldY = b2.dot(pixEyePos) + fromPoint.y;
+  double pixWorldZ = b3.dot(pixEyePos) + fromPoint.z;
 
   Vector3 pixWorldPos = Vector3(pixWorldX, pixWorldY, pixWorldZ);
 
@@ -28,12 +28,13 @@ Ray Camera::computeRayAt(int i, int j, int N) {
 
   return Ray(pixWorldPos, rayDir);
 }
+
 const Vector3 &Camera::getFromPoint() const {
   return fromPoint;
 }
 
 void Camera::moveRight() {
-  Vector3 dir(0.1f, 0 , 0);
+  Vector3 dir(0.1, 0 , 0);
 
   fromPoint += dir;
 
@@ -41,7 +42,7 @@ void Camera::moveRight() {
 }
 
 void Camera::moveLeft() {
-  Vector3 dir(-0.1f, 0 , 0);
+  Vector3 dir(-0.1, 0 , 0);
 
   fromPoint += dir;
 
@@ -49,14 +50,14 @@ void Camera::moveLeft() {
 }
 
 void Camera::moveUp() {
-  Vector3 dir(0, 0.1f , 0);
+  Vector3 dir(0, 0.1 , 0);
 
   fromPoint += dir;
 
   computeBasisVectors();
 }
 void Camera::moveDown() {
-  Vector3 dir(0, -0.1f , 0);
+  Vector3 dir(0, -0.1 , 0);
 
   fromPoint += dir;
