+++ Intersection.cpp
@@ -3,4 +3,4 @@
 //
 
 #include "Intersection.h"
-Intersection::Intersection(const Sphere &sphere, float t) : sphere(sphere), t(t) {}
+Intersection::Intersection(const Sphere &sphere, double t) : sphere(sphere), t(t) {}
